Many minor fixes. Fix reaction rolls, add any screenshoots, fix generate starting spells, bull strenght use percent if already 18.

+++ eob2/math.h
@@ -8,5 +8,5 @@ template<class T> inline void iswap(T& a, T& b) { T i = a; a = b; b = i; }
 int isqrt(int num);
 float sqrt(const float x);
 
-#define maptbl(t, v) (t[imax((size_t)0, imin((size_t)v, sizeof(t)/sizeof(t[0])-1))])
+#define maptbl(t, v) (t[imax((int)0, imin((int)v, (int)(sizeof(t)/sizeof(t[0])-1)))])
 #define FG(v) (1<<v)
